% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilogit.R
\name{multilogit}
\alias{multilogit}
\title{Multinomial Logistic Regression using Data Augmentation}
\usage{
multilogit(
  Y,
  X,
  n_sample = 1000,
  n_burn = 200,
  n_sigma_check = 20,
  step_size = 0.1,
  prior = "flat",
  prior_mean = NULL,
  prior_var = NULL,
  reference_cat = NULL,
  probs = FALSE,
  progress = TRUE
)
}
\arguments{
\item{Y}{An N by C numeric matrix where the ith row is a set of
indicators for observation i of N total observations giving which
of the C categories the observation is classified into.}

\item{X}{An N by P numeric matrix where N is the total number of
observations and P is the total number of predictor variables the ith row
gives the values of the predictor variables for the ith outcome observation.}

\item{n_sample}{positive integer giving the number of samples to draw as
output after burn-in.}

\item{n_burn}{non-negative integer giving the number of samples of burn-in
before the chain output is saved.}

\item{n_sigma_check}{positive integer giving the period for the number of
samples on which to automatically tune the step size of the random walk.}

\item{step_size}{finite positive standard deviation for the random-walk
proposal distribution.}

\item{prior}{character with either values "flat" or "normal"}

\item{prior_mean}{a vector of length equal to P the number of predictors.
Giving the mean for the normal prior on the coefficient vector. Only use if
prior = "normal".}

\item{prior_var}{a P by P matrix giving the covariance matrix of the prior
coefficients. Note that covariance structure is assumed to be constant across
categories. Must be a positive-definite matrix. Only use with prior =
"normal".}

\item{reference_cat}{Either NULL or an integer between 1 and C where C is
the total number of categories. If left NULL, coefficients for all
categories will be generated, leading to a loss of identifiability in the
coefficients. If an integer, all of the coefficients for that category will
be held constant at 0.}

\item{probs}{logical; if TRUE, return fitted category probabilities for each
retained draw.}

\item{progress}{logical; if TRUE, print periodic progress updates.}
}
\value{
List object containing posterior_coef, the chain of coefficient
values as an P by C by n_sample array. And posterior_prob a N by C by
n_sample array containing the calculated probabilities of the observations
being classified into each of the C categories.
}
\description{
This function implements a data augmentation method for
multinomial logistic regression using MCMC. The data augmentation method is
outlined in a paper by Jared D. Fisher and Kyle R. McEvoy titled
"Bayesian Multinomial Logistic Regression for Numerous Categories",
currently a work in progress. The sampler uses a Random-Walk Metropolis
algorithm for choosing proposals and acceptances.
}
\examples{
Y <- matrix(0, nrow = 150, ncol = 3)
Y[1:50, 1] <- 1
Y[51:100, 2] <- 1
Y[101:150, 3] <- 1
X <- cbind(1, iris[ , 1:4])
X <- as.matrix(X)
out <- multilogit(Y, X, n_sample = 3000, n_burn = 1500, n_sigma_check = 20,
 step_size = 0.1, prior = "normal", reference_cat = 1)
 
out_2 <- multilogit(Y, X, n_sample = 2000, n_burn = 500, prior = "flat")

}
