% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilogit_holmesheld.R
\name{multilogit_holmesheld}
\alias{multilogit_holmesheld}
\title{Multinomial Logistic Regression using the Holmes-Held Method}
\usage{
multilogit_holmesheld(
  Y,
  X,
  v = diag(10, nrow = ncol(X)),
  n_sample = 1000L,
  n_burn = 200L
)
}
\arguments{
\item{Y}{An N by C numeric matrix where the ith row is a set of
indicators for observation i of N total observations giving which
of the C categories the observation is classified into.}

\item{X}{An N by P numeric matrix where the ith row gives
the values of the predictor variables for the ith outcome observation.}

\item{v}{a P by P numeric matrix giving the covariance matrix of coefficients.
The function only accepts one matrix for all categories.}

\item{n_sample}{positive integer giving the number of samples to draw as
output after burn-in.}

\item{n_burn}{non-negative integer giving the number of samples of burn-in
before the chain output is saved.}
}
\value{
List object containing posterior_coef, the chain of coefficient
values as an P by C by n_sample array. And posterior_prob a N by C by
n_sample array containing the calculated probabilities of the observations
being classified into each of the C categories.
}
\description{
This is an R wrapper to the function multilogit_holmesheld_C.
This function implements the Holmes-Held method for
multinomial logistic regression described in their 2006 paper
'Bayesian auxiliary variable models for binary and multinomial regression'
in Bayesian Analysis. The C++ code was written using the pseudo-code in
this paper as a template.
}
\examples{
Y <- matrix(0, nrow = 150, ncol = 3)
Y[1:50, 1] <- 1
Y[51:100, 2] <- 1
Y[101:150, 3] <- 1
X <- cbind(1, iris[ , 1:4])
X <- as.matrix(X)
v <- diag(10, ncol(X))
out <- multilogit_holmesheld(Y, X, v, n_sample = 2000, n_burn = 1000)


}
